# openapi/railkeeper.yaml
openapi: 3.1.0
info:
  title: RailKeeper2 API
  version: 0.1.0
servers:
  - url: /api/v1
tags:
  - name: Setup
  - name: Auth
  - name: Vehicles
  - name: MasterData
  - name: ArticleSearch
  - name: Backup
  - name: Audit
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rk_session
  schemas:
    Problem:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
    SetupStatus:
      type: object
      required: [setupRequired]
      properties:
        setupRequired:
          type: boolean
    CreateAdminRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 12
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    Session:
      type: object
      required: [username, roles, csrfToken]
      properties:
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        csrfToken:
          type: string
    Vehicle:
      type: object
      required: [id, inventoryNumber, manufacturer, name, gauge, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        inventoryNumber:
          type: string
        manufacturer:
          type: string
        articleNumber:
          type: string
        name:
          type: string
        gauge:
          type: string
        epoch:
          type: string
        railwayCompany:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateVehicleRequest:
      type: object
      required: [manufacturer, name, gauge]
      properties:
        inventoryNumber:
          type: string
        manufacturer:
          type: string
        articleNumber:
          type: string
        name:
          type: string
        gauge:
          type: string
        epoch:
          type: string
        railwayCompany:
          type: string
        category:
          type: string
paths:
  /setup/status:
    get:
      tags: [Setup]
      security: []
      responses:
        "200":
          description: Setup state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"
  /setup/admin:
    post:
      tags: [Setup]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminRequest"
      responses:
        "201":
          description: Admin user created
        "400":
          description: Weak input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Setup already completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/session:
    get:
      tags: [Auth]
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/login:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/logout:
    post:
      tags: [Auth]
      responses:
        "204":
          description: Logged out
  /vehicles:
    get:
      tags: [Vehicles]
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vehicle list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      tags: [Vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          description: Invalid vehicle data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /article-search:
    get:
      tags: [ArticleSearch]
      parameters:
        - name: manufacturer
          in: query
          schema:
            type: string
        - name: articleNumber
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Article search results